remove panic on unrecognized change
tl;dr

panics are super scary, probably don't need to panic just because
we barfed on one change. simply logging it, and returning should
be fine.

---

Motivation:

We were looking through the source of jsoncdc, and had noticed there was
a panic inside `append_change`. We couldn't think of a reason to panic
here (although we don't have as much insight I'm sure). Talking it over
we decided to switch to simply logging the unrecognized change action
instead.

We don't see a problem with replacing this panic (and all of the tests
still pass), however if you want we can change this PR to use a
compile time feature that isn't on by default.

Docker Changes:

I was unable to build the dockerfile when attempting to run tests.
Specifically it failed building `rustfmt`, by adding in gcc-multilib
I was able to build the dockerfile so I've added that into my change too.

util/travis Changes:

I was unable to get travis to build correctly. so I made two changes
The first was changing substitution of `"$LD_LIBRARY_PATH"` to:
`"${LD_LIBRARY_PATH:-}"` which sets the string to empty if it's not defined
since if it's not defined the script errors due to: `set -o nounset`

I also was unable to get this to build onto mac, since postgis had developed
a dependency on outdated postgres. this now forces the uninstall of the old
postgres in order to make sure we can get a current one.